[Improve][Connector-V2][Jdbc-Source] Support for Decimal types as splict keys  (#4634)
* [Improve][Connector-V2][Jdbc-Source]Support Compatible Mysql bigint(20) used as a partition_column #4634

+++ seatunnel-connectors-v2/connector-jdbc/src/main/java/org/apache/seatunnel/connectors/seatunnel/jdbc/config/JdbcSourceConfig.java
@@ -23,6 +23,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Data
@@ -33,8 +34,8 @@ public class JdbcSourceConfig implements Serializable {
     private JdbcConnectionConfig jdbcConnectionConfig;
     public String query;
     private String partitionColumn;
-    private Long partitionUpperBound;
-    private Long partitionLowerBound;
+    private BigDecimal partitionUpperBound;
+    private BigDecimal partitionLowerBound;
     private int fetchSize;
     private Integer partitionNumber;
 
@@ -60,11 +61,11 @@ public class JdbcSourceConfig implements Serializable {
         return Optional.ofNullable(partitionColumn);
     }
 
-    public Optional<Long> getPartitionUpperBound() {
+    public Optional<BigDecimal> getPartitionUpperBound() {
         return Optional.ofNullable(partitionUpperBound);
     }
 
-    public Optional<Long> getPartitionLowerBound() {
+    public Optional<BigDecimal> getPartitionLowerBound() {
         return Optional.ofNullable(partitionLowerBound);
     }
